+++ modeling.py
@@ -1,0 +1,25 @@
+from sklearn.model_selection import GridSearchCV
+from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
+
+
+def hyperparameter_tuning(model, param_grid, X_train, y_train):
+    """Perform grid search to tune model hyperparameters."""
+    grid_search = GridSearchCV(
+        estimator=model,
+        param_grid=param_grid,
+        cv=5,
+        scoring='neg_mean_squared_error',
+        n_jobs=-1
+    )
+    grid_search.fit(X_train, y_train)
+    return grid_search.best_estimator_
+
+
+def evaluate_model(model, X_test, y_test):
+    """Return predictions and common regression metrics."""
+    y_pred = model.predict(X_test).flatten()
+    mse = mean_squared_error(y_test, y_pred)
+    mae = mean_absolute_error(y_test, y_pred)
+    r2 = r2_score(y_test, y_pred)
+    return y_pred, mse, mae, r2
+
